refactor: mvc 패턴 구조 수정
